add optional fields to torrent files
these are somewhat common fields included in the info dictionary, so
it's probably worth including them. In this case I needed to do so
because otherwise my infohash computation is wrong because i decode and
reencode instead of taking a slice of the original file.

i can actually use these for checks later if i so desire